plugins/social_network/social_network_impl/firebase_config.py
import os
import logging
from dotenv import load_dotenv
import firebase_admin
from firebase_admin import credentials, auth
import pyrebase

# Load environment variables
load_dotenv()

# Firebase configuration
firebase_config = {
    "apiKey": os.getenv('FIREBASE_WEB_API_KEY'),
    "authDomain": f"{os.getenv('FIREBASE_PROJECT_ID')}.firebaseapp.com",
    "projectId": os.getenv('FIREBASE_PROJECT_ID'),
    "storageBucket": f"{os.getenv('FIREBASE_PROJECT_ID')}.appspot.com",
    "messagingSenderId": os.getenv('FIREBASE_PROJECT_NUMBER'),
    "databaseURL": ""  # Optional, leave blank if not using Realtime Database
}

# Initialize Firebase Admin
try:
    firebase_admin.initialize_app()
except ValueError:
    # App already initialized
    pass

# Pyrebase configuration
try:
    pyrebase_app = pyrebase.initialize_app(firebase_config)
    pyrebase_auth = pyrebase_app.auth()
except Exception as e:
    logging.error(f"Pyrebase initialization error: {e}")
    pyrebase_auth = None

def initialize_firebase():
    """
    Initialize Firebase configuration
    """
    logging.info("Firebase configuration loaded")
    return firebase_config

def google_sign_in():
    """
    Initiate Google Sign-In process
    
    Returns:
        dict: User information if sign-in is successful
        None: If sign-in fails
    """
    try:
        # Create a Google Auth Provider
        google_provider = "google.com"
        
        # This is a placeholder. In a real implementation, 
        # you would use Firebase Authentication UI or a web-based flow
        logging.info("Attempting Google Sign-In")
        
        # Simulate a Google Sign-In (this is just a mock)
        # In a real app, this would involve actual OAuth flow
        return {
            "success": True,
            "message": "Google Sign-In simulated successfully",
            "user": {
                "email": "example@gmail.com",
                "displayName": "Example User"
            }
        }
    except Exception as e:
        logging.error(f"Google Sign-In error: {e}")
        return {
            "success": False,
            "message": str(e)
        }

def create_collections():
    """
    Placeholder for creating initial collections
    """
    logging.info("Collections creation placeholder")
    print("Firebase collections initialized (placeholder).")
