plugins/social_network/social_network_impl/requirements.txt
PyQt5==5.15.7
python-dotenv==0.21.0
firebase-admin==6.2.0
pyrebase4==4.7.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.28.2
bcrypt==4.0.1
